backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pandas==2.2.3
python-dotenv==1.0.1
python-multipart==0.0.20
google-generativeai==0.8.4
pydantic==2.10.4
aiosqlite==0.20.0
sqlparse==0.5.3
numpy==2.2.1
